Fix the LOD combining Ask for input of what map chunk to process at what detail level
Convert everything to the game's world-space coordinates, optionally offsetting them by half of the game world so that it's centered in Blender's world. This also allowed me to get rid of the magic numbers for the terrain object's X/Y scale values. Magic Z scale remains for now, but it's now baked into the vertex coordinates instead of object scale.

Call shade_smooth() so it looks noicer.